backend/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-sqlalchemy>=3.1
python-dotenv>=1.0
sqlalchemy>=2.0

# AI / LLM
langchain>=0.3
langchain-google-genai>=2.0
google-generativeai>=0.8

# Data processing
pandas>=2.1
scikit-learn>=1.5
numpy>=1.26

# Web scraping
beautifulsoup4>=4.12
requests>=2.31
lxml>=5.0

# Utilities
pydantic>=2.9
python-multipart>=0.0.12
werkzeug>=3.0

# Testing
pytest>=7.4
pytest-cov>=4.1
